split stats into its own file, add computeStatsAdhoc function
