+++ functions/package.json
@@ -14,7 +14,7 @@
     "@google-cloud/pubsub": "^1.1.5",
     "cors": "^2.8.5",
     "express": "^4.17.1",
-    "firebase-admin": "^8.6.1",
+    "firebase-admin": "^8.12.1",
     "firebase-functions": "^3.3.0",
     "gm": "^1.23.1",
     "json2csv": "^4.5.4",
@@ -44,7 +44,9 @@
   "engines": {
     "node": "10"
   },
-  "eslintIgnore": ["*.test.ts"],
+  "eslintIgnore": [
+    "*.test.ts"
+  ],
   "private": true,
   "main": "dist/index.js"
 }
