+++ functions/src/constants.ts
@@ -1,0 +1,1 @@
+export const STATS_TOPIC = "update-stats";
