+++ functions/src/firestore.ts
@@ -1,0 +1,7 @@
+import admin from "firebase-admin";
+
+admin.initializeApp();
+export const firestore = admin.firestore();
+export const auth = admin.auth();
+const settings = { timestampsInSnapshots: true };
+firestore.settings(settings);
