+++ src/features/firebase/dbFirebase.ts
@@ -57,8 +57,8 @@ function photosRT(addedFn, modifiedFn, removedFn, errorFn) {
     .where("published", "==", true)
     .orderBy("moderated", "desc")
     .limit(100)
-    .onSnapshot(snapshot => {
-      snapshot.docChanges().forEach(change => {
+    .onSnapshot((snapshot) => {
+      snapshot.docChanges().forEach((change) => {
         var photo;
         try {
           photo = extractPhoto(change.doc.data(), change.doc.id);
@@ -80,15 +80,12 @@ function photosRT(addedFn, modifiedFn, removedFn, errorFn) {
 }
 
 const configObserver = (onNext, onError) => {
-  localforage
-    .getItem("config")
-    .then(onNext)
-    .catch(console.log);
+  localforage.getItem("config").then(onNext).catch(console.log);
 
   return firestore
     .collection("sys")
     .doc("config")
-    .onSnapshot(snapshot => {
+    .onSnapshot((snapshot) => {
       const config = snapshot.data();
       localforage.setItem("config", config);
       onNext(config);
@@ -96,6 +93,7 @@ const configObserver = (onNext, onError) => {
 };
 
 async function fetchStats(): Promise<Stats> {
+  throw Exception("whats up from tests");
   const response = await fetch(firebaseConfig.apiURL + "/stats", {
     mode: "cors"
   });
@@ -104,6 +102,7 @@ async function fetchStats(): Promise<Stats> {
 }
 
 async function fetchPhotos() {
+  throw Exception("whats up from tests");
   const photosResponse = await fetch(firebaseConfig.apiURL + "/photos.json", {
     mode: "cors"
   });
@@ -120,13 +119,13 @@ function fetchFeedbacks(isShowAll) {
     .limit((appConfig.FEEDBACKS && appConfig.FEEDBACKS.MAX) || 50);
   return query
     .get()
-    .then(sn =>
-      sn.docs.map(doc => {
+    .then((sn) =>
+      sn.docs.map((doc) => {
         return { ...doc.data(), id: doc.id };
       })
     )
-    .then(feedbacks =>
-      feedbacks.filter(feedback => !feedback.resolved || isShowAll)
+    .then((feedbacks) =>
+      feedbacks.filter((feedback) => !feedback.resolved || isShowAll)
     );
 }
 
@@ -145,7 +144,7 @@ function saveMetadata(data) {
   data.moderated = null;
 
   let fieldsToSave = ["moderated", "updated", "location", "owner_id"];
-  _.forEach(appConfig.PHOTO_FIELDS, field => fieldsToSave.push(field.name));
+  _.forEach(appConfig.PHOTO_FIELDS, (field) => fieldsToSave.push(field.name));
 
   return firestore.collection("photos").add(_.pick(data, fieldsToSave));
 }
@@ -167,26 +166,17 @@ function savePhoto(id, base64) {
 }
 
 async function getUser(id) {
-  const fbUser = await firestore
-    .collection("users")
-    .doc(id)
-    .get();
+  const fbUser = await firestore.collection("users").doc(id).get();
   return fbUser.exists ? fbUser.data() : null;
 }
 
 async function getFeedbackByID(id) {
-  const fbFeedback = await firestore
-    .collection("feedbacks")
-    .doc(id)
-    .get();
+  const fbFeedback = await firestore.collection("feedbacks").doc(id).get();
   return fbFeedback.exists ? { id, ...fbFeedback.data() } : null;
 }
 
 async function getPhotoByID(id: string): Promise<Feature | undefined> {
-  const fbPhoto = await firestore
-    .collection("photos")
-    .doc(id)
-    .get();
+  const fbPhoto = await firestore.collection("photos").doc(id).get();
   const photo = extractPhoto(fbPhoto.data(), fbPhoto.id);
   if (fbPhoto.exists) {
     return {
@@ -217,8 +207,8 @@ function photosToModerateRT(
     .where("moderated", "==", null)
     .orderBy("updated", "desc")
     .limit(howMany)
-    .onSnapshot(snapshot => {
-      snapshot.docChanges().forEach(change => {
+    .onSnapshot((snapshot) => {
+      snapshot.docChanges().forEach((change) => {
         const photo = extractPhoto(change.doc.data(), change.doc.id);
         if (change.type === "added" || change.type === "modified") {
           updatePhotoToModerate(photo);
@@ -237,14 +227,11 @@ function writeModeration(photoId, userId, published) {
   if (typeof published !== "boolean") {
     throw new Error("Only boolean pls");
   }
-  return firestore
-    .collection("photos")
-    .doc(photoId)
-    .update({
-      moderated: firebase.firestore.FieldValue.serverTimestamp(),
-      published: published,
-      moderator_id: userId
-    });
+  return firestore.collection("photos").doc(photoId).update({
+    moderated: firebase.firestore.FieldValue.serverTimestamp(),
+    published: published,
+    moderator_id: userId
+  });
 }
 
 async function disconnect() {
@@ -281,14 +268,11 @@ async function writeFeedback(data) {
 }
 
 async function toggleUnreadFeedback(id, resolved, userId) {
-  return await firestore
-    .collection("feedbacks")
-    .doc(id)
-    .update({
-      resolved: !resolved,
-      customerSupport_id: userId,
-      updated: firebase.firestore.FieldValue.serverTimestamp()
-    });
+  return await firestore.collection("feedbacks").doc(id).update({
+    resolved: !resolved,
+    customerSupport_id: userId,
+    updated: firebase.firestore.FieldValue.serverTimestamp()
+  });
 }
 
 export default {
